Changed MovementRange to floats
Makes for nicer(?) circles? Can't decide if I like 2.5 better than 2 or not

+++ Assets/ScriptableObjects/TileData.cs
@@ -7,5 +7,5 @@ public class TileData : ScriptableObject
     public TileBase[] tiles;
 
     public bool navigable;
-    public int movementRange;
+    public float movementRange;
 }
